Code cleanup but the catch is stuff broke pls fix
