+++ gui/custom_widgets.py
@@ -43,6 +43,7 @@ class ResizeHandle(QGraphicsRectItem):
 class EditableGraphicsItem(QObject):
     """A wrapper class that adds editing capabilities to QGraphicsItems"""
     itemChanged = Signal(QGraphicsItem)
+    itemIsChanging = Signal(QGraphicsItem)
     transformChanged = Signal(QTransform)
     itemSelected = Signal(str)
     
@@ -156,20 +157,14 @@ class EditableGraphicsItem(QObject):
         """Update the position of bounding box and handles after item moves"""
         try:
             if self.isItemDeleted(self.item):
-                # If the wrapped item itself is deleted, remove all adornments and stop.
                 if self.boundingBoxItem or self.handles or self.rotationHandle:
                     print(f"Main item {id(self.item)} deleted. Removing its bounding box and handles.")
                 self.removeBoundingBox()
                 return
 
-            # Ensure our bounding box graphics item is still valid and exists
             if not self.boundingBoxItem or self.isItemDeleted(self.boundingBoxItem):
                 if self.boundingBoxItem and self.isItemDeleted(self.boundingBoxItem):
-                    # It was deleted by Qt, so nullify our reference.
                     self.boundingBoxItem = None
-                # Cannot update if the box item itself is gone.
-                # setupBoundingBox would be needed to recreate it if desired.
-                # print(f"BoundingBoxItem for item {id(self.item)} is None or deleted. Cannot update.")
                 return 
 
             rect = self.item.boundingRect()
@@ -186,14 +181,13 @@ class EditableGraphicsItem(QObject):
                 HandleType.Left: QPointF(rect.left(), rect.center().y()),
             }
             
-            active_handle_types = list(self.handles.keys()) # Iterate over a copy of keys
+            active_handle_types = list(self.handles.keys())
             for handle_type in active_handle_types:
                 handle = self.handles.get(handle_type)
                 if handle:
                     if not self.isItemDeleted(handle):
                         handle.setPos(handle_positions[handle_type])
                     else:
-                        # Handle was deleted by Qt, remove our reference
                         del self.handles[handle_type]
             
             if self.rotationHandle:
@@ -201,7 +195,6 @@ class EditableGraphicsItem(QObject):
                     rotation_pos = QPointF(rect.center().x(), rect.top() - self.rotateHandleOffset)
                     self.rotationHandle.setPos(rotation_pos)
                 else:
-                    # Rotation handle was deleted by Qt
                     self.rotationHandle = None
                 
         except (RuntimeError, AttributeError) as e:
@@ -238,7 +231,7 @@ class EditableGraphicsItem(QObject):
     
     def handleEditOperation(self, new_pos):
         """Handle move, resize, or rotate operations"""
-        if not self.isEditing or not self.currentHandle:
+        if not self.isEditing or self.currentHandle is None:
             return
             
         if self.currentHandle == HandleType.Rotation:
@@ -247,7 +240,7 @@ class EditableGraphicsItem(QObject):
             self.handleResize(new_pos)
         
         self.updateBoundingBox()
-        self.itemChanged.emit(self.item)
+        self.itemIsChanging.emit(self.item)
     
     def handleRotation(self, mouse_pos):
         """Handle rotation operation"""
@@ -257,497 +250,351 @@ class EditableGraphicsItem(QObject):
         dy = mouse_pos.y() - center.y()
         current_angle = math.degrees(math.atan2(dy, dx))
         
-        angle_diff = current_angle - self.initialAngle
+        rotation_delta = current_angle - self.initialAngle
         
-        center_local = self.initialBoundingRect.center()
+        transform = QTransform()
+        transform.translate(self.initialItemPos.x(), self.initialItemPos.y())
+        transform.rotate(self.initialRotation + rotation_delta)
+        transform.translate(-self.initialItemPos.x(), -self.initialItemPos.y())
+        
+        self.item.setTransform(self.initialItemTransform * transform)
+        
+    def handleResize(self, mouse_pos_scene):
+        handle_type = self.currentHandle
+        
+        inverted_transform, _ = self.initialItemTransform.inverted()
+        mouse_pos_item = inverted_transform.map(mouse_pos_scene)
+        
+        initial_pos_item = inverted_transform.map(self.initialClickPos)
+        
+        delta = mouse_pos_item - initial_pos_item
+        
+        new_rect = QRectF(self.initialBoundingRect)
+        
+        if handle_type == HandleType.TopLeft:
+            new_rect.setTopLeft(new_rect.topLeft() + delta)
+        elif handle_type == HandleType.Top:
+            new_rect.setTop(new_rect.top() + delta.y())
+        elif handle_type == HandleType.TopRight:
+            new_rect.setTopRight(new_rect.topRight() + delta)
+        elif handle_type == HandleType.Right:
+            new_rect.setRight(new_rect.right() + delta.x())
+        elif handle_type == HandleType.BottomRight:
+            new_rect.setBottomRight(new_rect.bottomRight() + delta)
+        elif handle_type == HandleType.Bottom:
+            new_rect.setBottom(new_rect.bottom() + delta.y())
+        elif handle_type == HandleType.BottomLeft:
+            new_rect.setBottomLeft(new_rect.bottomLeft() + delta)
+        elif handle_type == HandleType.Left:
+            new_rect.setLeft(new_rect.left() + delta.x())
+        
+        # To prevent inverted rects
+        if new_rect.width() < 1: new_rect.setWidth(1)
+        if new_rect.height() < 1: new_rect.setHeight(1)
 
+        # This part requires the item to have a setRect or similar method.
+        # This will work for QGraphicsRectItem, but for QGraphicsPixmapItem etc.,
+        # we need to adjust scale, not rect.
+        
+        # For now, let's assume we can scale based on rect change
+        x_scale = new_rect.width() / self.initialBoundingRect.width() if self.initialBoundingRect.width() != 0 else 1
+        y_scale = new_rect.height() / self.initialBoundingRect.height() if self.initialBoundingRect.height() != 0 else 1
+        
         new_transform = QTransform(self.initialItemTransform)
-        new_transform.translate(center_local.x(), center_local.y())
-        new_transform.rotate(angle_diff)
-        new_transform.translate(-center_local.x(), -center_local.y())
+        
+        # We need to translate to origin, scale, and translate back
+        center = self.initialBoundingRect.center()
+        new_transform.translate(center.x(), center.y())
+        new_transform.scale(x_scale, y_scale)
+        new_transform.translate(-center.x(), -center.y())
         
         self.item.setTransform(new_transform)
-    
-    def handleResize(self, mouse_pos_scene):
-        """Handle resize operation"""
-        if not self.isEditing or not self.currentHandle or self.currentHandle == HandleType.Rotation:
-            return
 
-        current_mouse_local = self.item.mapFromScene(mouse_pos_scene)
-
-        new_local_rect = QRectF(self.originalRect)
-
-        if self.currentHandle == HandleType.TopLeft:
-            new_local_rect.setTopLeft(current_mouse_local)
-            new_local_rect.setBottomRight(self.originalRect.bottomRight()) 
-        elif self.currentHandle == HandleType.Top:
-            new_local_rect.setTop(current_mouse_local.y())
-            new_local_rect.setBottom(self.originalRect.bottom())
-        elif self.currentHandle == HandleType.TopRight:
-            new_local_rect.setTopRight(current_mouse_local)
-            new_local_rect.setBottomLeft(self.originalRect.bottomLeft()) 
-        elif self.currentHandle == HandleType.Right:
-            new_local_rect.setRight(current_mouse_local.x())
-            new_local_rect.setLeft(self.originalRect.left())   
-        elif self.currentHandle == HandleType.BottomRight:
-            new_local_rect.setBottomRight(current_mouse_local)
-            new_local_rect.setTopLeft(self.originalRect.topLeft())   
-        elif self.currentHandle == HandleType.Bottom:
-            new_local_rect.setBottom(current_mouse_local.y())
-            new_local_rect.setTop(self.originalRect.top())   
-        elif self.currentHandle == HandleType.BottomLeft:
-            new_local_rect.setBottomLeft(current_mouse_local)
-            new_local_rect.setTopRight(self.originalRect.topRight()) 
-        elif self.currentHandle == HandleType.Left:
-            new_local_rect.setLeft(current_mouse_local.x())
-            new_local_rect.setRight(self.originalRect.right())
-
-        new_local_rect = new_local_rect.normalized()
-        
-        # minimum size so you can see it I guess (change if needed)
-        min_size = 50.0
-
-        if new_local_rect.width() < min_size:
-            if self.currentHandle in [HandleType.Left, HandleType.TopLeft, HandleType.BottomLeft]:
-                new_local_rect.setLeft(new_local_rect.right() - min_size)
-            elif self.currentHandle in [HandleType.Right, HandleType.TopRight, HandleType.BottomRight]:
-                new_local_rect.setRight(new_local_rect.left() + min_size)
-            else: 
-                delta_w = min_size - new_local_rect.width()
-                new_local_rect.adjust(-delta_w / 2, 0, delta_w / 2, 0)
-
-        if new_local_rect.height() < min_size:
-            if self.currentHandle in [HandleType.Top, HandleType.TopLeft, HandleType.TopRight]:
-                new_local_rect.setTop(new_local_rect.bottom() - min_size)
-            elif self.currentHandle in [HandleType.Bottom, HandleType.BottomLeft, HandleType.BottomRight]:
-                new_local_rect.setBottom(new_local_rect.top() + min_size)
-            else:
-                delta_h = min_size - new_local_rect.height()
-                new_local_rect.adjust(0, -delta_h / 2, 0, delta_h / 2)
-        
-        new_local_rect = new_local_rect.normalized()
-
-        if new_local_rect.width() < min_size:
-            new_local_rect.setWidth(min_size)
-        if new_local_rect.height() < min_size:
-            new_local_rect.setHeight(min_size)
-        
-        if hasattr(self.item, 'setRect'):
-            self.item.setRect(new_local_rect)
-        elif hasattr(self.item, 'path') and hasattr(self, 'original_path_at_drag_start'):
-            op_bounds = self.original_path_at_drag_start.boundingRect()
-            if op_bounds.width() == 0 or op_bounds.height() == 0: 
-                self.updateBoundingBox()
-                self.itemChanged.emit(self.item)
-                return
-
-            scale_x = new_local_rect.width() / op_bounds.width()
-            scale_y = new_local_rect.height() / op_bounds.height()
-
-            pivot = QPointF()
-            if self.currentHandle == HandleType.TopLeft: pivot = op_bounds.bottomRight()
-            elif self.currentHandle == HandleType.Top: pivot = QPointF(op_bounds.center().x(), op_bounds.bottom())
-            elif self.currentHandle == HandleType.TopRight: pivot = op_bounds.bottomLeft()
-            elif self.currentHandle == HandleType.Right: pivot = QPointF(op_bounds.left(), op_bounds.center().y())
-            elif self.currentHandle == HandleType.BottomRight: pivot = op_bounds.topLeft()
-            elif self.currentHandle == HandleType.Bottom: pivot = QPointF(op_bounds.center().x(), op_bounds.top())
-            elif self.currentHandle == HandleType.BottomLeft: pivot = op_bounds.topRight()
-            elif self.currentHandle == HandleType.Left: pivot = QPointF(op_bounds.right(), op_bounds.center().y())
-            
-            transform = QTransform().translate(pivot.x(), pivot.y()).scale(scale_x, scale_y).translate(-pivot.x(), -pivot.y())
-            scaled_path = transform.map(self.original_path_at_drag_start)
-            self.item.setPath(scaled_path)
-            
-        elif isinstance(self.item, QGraphicsEllipseItem):
-             self.item.setRect(new_local_rect)
-        else:
-            if self.originalRect.width() == 0 or self.originalRect.height() == 0: 
-                self.updateBoundingBox()
-                self.itemChanged.emit(self.item)
-                return
-
-            scale_x = new_local_rect.width() / self.originalRect.width()
-            scale_y = new_local_rect.height() / self.originalRect.height()
-
-            pivot_in_item_coords = QPointF()
-            if self.currentHandle == HandleType.TopLeft:    pivot_in_item_coords = self.originalRect.bottomRight()
-            elif self.currentHandle == HandleType.Top:       pivot_in_item_coords = QPointF(self.originalRect.center().x(), self.originalRect.bottom())
-            elif self.currentHandle == HandleType.TopRight:  pivot_in_item_coords = self.originalRect.bottomLeft()
-            elif self.currentHandle == HandleType.Right:     pivot_in_item_coords = QPointF(self.originalRect.left(), self.originalRect.center().y())
-            elif self.currentHandle == HandleType.BottomRight:pivot_in_item_coords = self.originalRect.topLeft()
-            elif self.currentHandle == HandleType.Bottom:    pivot_in_item_coords = QPointF(self.originalRect.center().x(), self.originalRect.top())
-            elif self.currentHandle == HandleType.BottomLeft:pivot_in_item_coords = self.originalRect.topRight()
-            elif self.currentHandle == HandleType.Left:      pivot_in_item_coords = QPointF(self.originalRect.right(), self.originalRect.center().y())
-            
-            new_transform = QTransform(self.initialItemTransform)
-            new_transform.translate(pivot_in_item_coords.x(), pivot_in_item_coords.y())
-            new_transform.scale(scale_x, scale_y)
-            new_transform.translate(-pivot_in_item_coords.x(), -pivot_in_item_coords.y())
-            self.item.setTransform(new_transform)
-
-        self.updateBoundingBox()
-        self.itemChanged.emit(self.item)
-    
     def endEdit(self):
-        """End current edit operation"""
+        """End edit operation"""
+        if self.isEditing:
+            self.itemChanged.emit(self.item)
         self.isEditing = False
         self.currentHandle = None
-        self.transformChanged.emit(self.item.transform())
-    
+
     def getItemRotation(self):
-        """Get current rotation angle of the item"""
-        transform = self.item.transform()
-        return math.degrees(math.atan2(transform.m12(), transform.m11()))
+        """Get the item's rotation from its transform"""
+        return self.item.rotation()
 
 class CheckerboardGraphicsScene(QGraphicsScene):
     itemSelectedOnCanvas = Signal(str)
     
     def __init__(self, parent=None):
         super(CheckerboardGraphicsScene, self).__init__(parent)
-        self.backgroundColor1 = QColor(200, 200, 200)
-        self.backgroundColor2 = QColor(230, 230, 230)
-        self.checkerboardSize = 20
+        self.color1 = QColor(220, 220, 220)
+        self.color2 = QColor(240, 240, 240)
+        self.tileSize = 20
         self.editMode = False
-        self.currentEditableItem = None
-        self.activeHandle = None
-        self.isDragging = False
-        self.lastMousePos = QPointF()
         self.editableItems = {}
+        self.currentEditableItem = None
+        self.dragInProgress = False
+        self.lastMousePos = QPointF()
 
     def setBackgroundColor(self, color1, color2):
-        """Set the background colors for the checkerboard pattern"""
-        self.backgroundColor1 = color1
-        self.backgroundColor2 = color2
+        self.color1 = color1
+        self.color2 = color2
         self.update()
 
     def drawBackground(self, painter, rect):
-        super(CheckerboardGraphicsScene, self).drawBackground(painter, rect)
-        
         painter.save()
-        painter.fillRect(rect, self.backgroundColor1)
+        painter.setPen(Qt.NoPen)
         
-        left = int(rect.left()) - (int(rect.left()) % self.checkerboardSize)
-        top = int(rect.top()) - (int(rect.top()) % self.checkerboardSize)
-        
-        for x in range(left, int(rect.right()), self.checkerboardSize):
-            for y in range(top, int(rect.bottom()), self.checkerboardSize):
-                if (x // self.checkerboardSize + y // self.checkerboardSize) % 2 == 0:
-                    painter.fillRect(x, y, self.checkerboardSize, self.checkerboardSize, self.backgroundColor2)
+        for y in range(int(rect.top()), int(rect.bottom()), self.tileSize):
+            for x in range(int(rect.left()), int(rect.right()), self.tileSize):
+                if (x // self.tileSize + y // self.tileSize) % 2 == 0:
+                    painter.setBrush(self.color1)
+                else:
+                    painter.setBrush(self.color2)
+                painter.drawRect(x, y, self.tileSize, self.tileSize)
         
         painter.restore()
-    
+
     def setEditMode(self, enabled):
-        """Enable or disable edit mode"""
         self.editMode = enabled
-        
-        if enabled:
-            items_to_cleanup = []
-            for item, editable_item in self.editableItems.items():
-                try:
-                    if not self.isItemDeleted(item):
-                        editable_item.setupBoundingBox()
-                    else:
-                        items_to_cleanup.append(item)
-                except (RuntimeError, AttributeError):
-                    items_to_cleanup.append(item)
-            
-            for item in items_to_cleanup:
-                if item in self.editableItems:
-                    del self.editableItems[item]
-        else:
+        if not enabled:
             self.deselectAllItems()
-            self.currentEditableItem = None
+            for item in self.items():
+                item.setFlag(QGraphicsItem.ItemIsMovable, False)
+                item.setFlag(QGraphicsItem.ItemIsSelectable, False)
+        else:
+            for item in self.items():
+                if item.data(1) == "Layer":
+                    item.setFlag(QGraphicsItem.ItemIsMovable, True)
+                    item.setFlag(QGraphicsItem.ItemIsSelectable, True)
 
     def makeItemEditable(self, item):
-        """Make a graphics item editable"""
-        if item not in self.editableItems:
-            editable_item = EditableGraphicsItem(item, self)
-            self.editableItems[item] = editable_item
-            
-            editable_item.itemChanged.connect(self.onItemChanged)
-            editable_item.itemSelected.connect(self.itemSelectedOnCanvas.emit)
-            
-            return editable_item
-        return self.editableItems[item]
+        if self.isItemDeleted(item):
+            return None
+        item_id = id(item)
+        if item_id not in self.editableItems:
+            editable = EditableGraphicsItem(item)
+            self.editableItems[item_id] = editable
+        return self.editableItems[item_id]
 
     def mousePressEvent(self, event):
-        """Handle mouse press events for editing"""
         if not self.editMode:
             super(CheckerboardGraphicsScene, self).mousePressEvent(event)
             return
             
-        item = self.itemAt(event.scenePos(), QTransform())
-        print(f"Mouse press on item: {item}")
+        pos = event.scenePos()
+        items_at_pos = self.items(pos)
         
-        if item and isinstance(item, ResizeHandle):
-            parent_item = item.parentItem()
-            print(f"  Handle type: {item.handle_type}, Parent: {parent_item}")
-            if parent_item in self.editableItems:
-                editable_item = self.editableItems[parent_item]
-                editable_item.startEdit(item.handle_type, event.scenePos())
-                self.currentEditableItem = editable_item
-                self.isDragging = True
-                event.accept()
-                return
-        elif item and item in self.editableItems:
-            self.selectItem(item)
-            super(CheckerboardGraphicsScene, self).mousePressEvent(event)
-            return
+        # Check if a handle was clicked first
+        for item in items_at_pos:
+            if isinstance(item, ResizeHandle):
+                # Look up the EditableGraphicsItem wrapper for the handle's parent
+                editable_parent = self.editableItems.get(id(item.parentItem()))
+                if editable_parent:
+                    self.currentEditableItem = editable_parent
+                    self.dragInProgress = True
+                    # Correctly start the edit operation on the wrapper
+                    self.currentEditableItem.startEdit(item.handle_type, pos)
+                    event.accept()
+                    return
+
+        # If not a handle, proceed with normal selection/move
+        top_item = self.itemAt(pos, QTransform())
+        if top_item and top_item.data(1) == "Layer":
+            self.selectItem(top_item)
+            self.dragInProgress = False
         else:
             self.deselectAllItems()
-            
+        
         super(CheckerboardGraphicsScene, self).mousePressEvent(event)
 
     def mouseMoveEvent(self, event):
-        """Handle mouse move events for editing"""
-        if self.editMode and self.isDragging and self.currentEditableItem:
-            print(f"Dragging in scene: pos={event.scenePos()}, dragging={self.isDragging}")
+        if self.editMode and self.currentEditableItem and self.dragInProgress:
             self.currentEditableItem.handleEditOperation(event.scenePos())
             event.accept()
-        else:
-            super(CheckerboardGraphicsScene, self).mouseMoveEvent(event)
-            
-            if self.editMode and self.currentEditableItem:
-                self.currentEditableItem.updateBoundingBox()
+            return
+        super(CheckerboardGraphicsScene, self).mouseMoveEvent(event)
 
     def mouseReleaseEvent(self, event):
-        """Handle mouse release events for editing"""
-        if self.editMode and self.isDragging and self.currentEditableItem:
-            self.currentEditableItem.endEdit()
-            self.isDragging = False
-            event.accept()
-            print(f"Edit ended, isDragging={self.isDragging}")
-        else:
-            super(CheckerboardGraphicsScene, self).mouseReleaseEvent(event)
-        
         if self.editMode and self.currentEditableItem:
+            if self.dragInProgress:
+                 self.currentEditableItem.endEdit()
+            # After finishing an edit or move, ensure the bounding box is up to date
             self.currentEditableItem.updateBoundingBox()
-    
+            
+        self.dragInProgress = False
+        super(CheckerboardGraphicsScene, self).mouseReleaseEvent(event)
+
     def selectItem(self, item):
-        """Select an item for editing"""
-        self.deselectAllItems()
+        if self.isItemDeleted(item):
+            self.currentEditableItem = None
+            return
+            
+        self.deselectAllItems(item_to_keep=item)
         
-        if item in self.editableItems and not self.isItemDeleted(item):
-            try:
-                editable_item = self.editableItems[item]
-                editable_item.setupBoundingBox()
-                self.currentEditableItem = editable_item
-                self.itemSelectedOnCanvas.emit(str(id(item)))
-            except (RuntimeError, AttributeError):
-                if item in self.editableItems:
-                    del self.editableItems[item]
-    
-    def deselectAllItems(self):
-        """Deselect all items"""
-        items_to_cleanup = []
-        for item, editable_item in self.editableItems.items():
-            try:
-                if not self.isItemDeleted(item):
-                    editable_item.removeBoundingBox()
-                else:
-                    items_to_cleanup.append(item)
-            except (RuntimeError, AttributeError):
-                items_to_cleanup.append(item)
+        editable = self.makeItemEditable(item)
+        if editable:
+            self.currentEditableItem = editable
+            editable.setupBoundingBox()
+            self.itemSelectedOnCanvas.emit(item.data(0))
         
-        for item in items_to_cleanup:
-            if item in self.editableItems:
-                del self.editableItems[item]
-                
-        self.currentEditableItem = None
-    
+        item.setSelected(True)
+
+    def deselectAllItems(self, item_to_keep=None):
+        if self.currentEditableItem and (not item_to_keep or self.currentEditableItem.item != item_to_keep):
+            self.currentEditableItem.removeBoundingBox()
+            self.currentEditableItem = None
+            
+        for item in self.items():
+            if item != item_to_keep:
+                item.setSelected(False)
+
     def isItemDeleted(self, item):
-        """Check if a QGraphicsItem has been deleted by Qt"""
         try:
             _ = item.pos()
             return False
-        except (RuntimeError, AttributeError):
+        except RuntimeError:
             return True
-    
+
     def onItemChanged(self, item):
-        """Handle item change events"""
-        pass
+        editable = self.editableItems.get(id(item))
+        if editable:
+            editable.updateBoundingBox()
 
 class CustomGraphicsView(QGraphicsView):
     def __init__(self, parent=None, min_zoom=0.05, max_zoom=10.0):
         super(CustomGraphicsView, self).__init__(parent)
-        self.min_zoom = min_zoom
-        self.max_zoom = max_zoom
-        self.current_zoom_level = 1.0
         self.setRenderHint(QPainter.Antialiasing)
         self.setRenderHint(QPainter.SmoothPixmapTransform)
         self.setTransformationAnchor(QGraphicsView.AnchorUnderMouse)
         self.setResizeAnchor(QGraphicsView.AnchorUnderMouse)
-        self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
-        self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
         self.setDragMode(QGraphicsView.ScrollHandDrag)
+        self.setInteractive(True)
+        self.min_zoom = min_zoom
+        self.max_zoom = max_zoom
         
-        self.is_panning = False
-        self.last_pan_point = QPointF()
-        
-        self.editMode = False
-        
-        self.gesture_active = False
-        self.initial_pinch_distance = 0.0
-        self.initial_zoom_level = 1.0
-        
-        self.grabGesture(Qt.PinchGesture)
-        
-        self.setViewportUpdateMode(QGraphicsView.FullViewportUpdate)
+        self.isMacOS = platform.system() == "Darwin"
+        if self.isMacOS:
+            self.setOptimizationFlag(QGraphicsView.DontAdjustForAntialiasing, True)
+            self.setHorizontalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+            self.setVerticalScrollBarPolicy(Qt.ScrollBarAlwaysOff)
+            self.viewport().setAttribute(Qt.WA_AcceptTouchEvents, True)
+            self.grabGesture(Qt.PinchGesture)
 
     def setEditMode(self, enabled):
-        """Enable or disable edit mode"""
-        self.editMode = enabled
-        
         if enabled:
             self.setDragMode(QGraphicsView.NoDrag)
+            if hasattr(self.scene(), 'setEditMode'):
+                self.scene().setEditMode(True)
         else:
             self.setDragMode(QGraphicsView.ScrollHandDrag)
-        
-        if self.scene() and hasattr(self.scene(), 'setEditMode'):
-            self.scene().setEditMode(enabled)
+            if hasattr(self.scene(), 'setEditMode'):
+                self.scene().setEditMode(False)
 
     def wheelEvent(self, event):
-        if not self.scene() or not self.scene().items():
-            event.accept()
-            return
-            
-        if hasattr(event, 'source') and event.source() == Qt.MouseEventSynthesizedBySystem:
-            if event.modifiers() == Qt.ControlModifier:
-                self.handleZoom(event.angleDelta().y())
-            else:
-                self.horizontalScrollBar().setValue(self.horizontalScrollBar().value() - event.angleDelta().x())
-                self.verticalScrollBar().setValue(self.verticalScrollBar().value() - event.angleDelta().y())
-        else:
-            if event.modifiers() == Qt.ControlModifier:
-                self.handleZoom(event.angleDelta().y())
-            else:
-                self.horizontalScrollBar().setValue(self.horizontalScrollBar().value() - event.angleDelta().x())
-                self.verticalScrollBar().setValue(self.verticalScrollBar().value() - event.angleDelta().y())
-            
-        event.accept()
-    
-    def handleZoom(self, delta):
-        zoom_factor = 1.0 + (delta / 1200.0)
-
-        new_zoom_level = self.current_zoom_level * zoom_factor
-        
-        if new_zoom_level < self.min_zoom:
-            new_zoom_level = self.min_zoom
-        elif new_zoom_level > self.max_zoom:
-            new_zoom_level = self.max_zoom
-        
-        zoom_change = new_zoom_level / self.current_zoom_level
-        self.scale(zoom_change, zoom_change)
-        self.current_zoom_level = new_zoom_level
-    
-    def mousePressEvent(self, event):
-        if not self.editMode and event.button() == Qt.LeftButton:
-            self.is_panning = True
-            self.last_pan_point = event.pos()
-            self.setCursor(Qt.ClosedHandCursor)
-            event.accept()
-        else:
-            super(CustomGraphicsView, self).mousePressEvent(event)
-
-    def mouseMoveEvent(self, event):
-        if not self.editMode and self.is_panning:
-            delta = event.pos() - self.last_pan_point
+        if self.isMacOS and (event.modifiers() == Qt.NoModifier or event.modifiers() == Qt.ShiftModifier):
+            # Scrolling gesture on macOS
+            delta = event.angleDelta()
             self.horizontalScrollBar().setValue(self.horizontalScrollBar().value() - delta.x())
             self.verticalScrollBar().setValue(self.verticalScrollBar().value() - delta.y())
-            self.last_pan_point = event.pos()
+            event.accept()
+        elif event.modifiers() & Qt.ControlModifier or (self.isMacOS and event.modifiers() & Qt.MetaModifier):
+            # Zooming
+            delta = event.angleDelta().y()
+            self.handleZoom(delta)
             event.accept()
         else:
-            super(CustomGraphicsView, self).mouseMoveEvent(event)
+            super(CustomGraphicsView, self).wheelEvent(event)
+
+    def handleZoom(self, delta):
+        zoom_factor = 1.15 if delta > 0 else 1 / 1.15
+        
+        current_transform = self.transform()
+        
+        if (delta > 0 and current_transform.m11() * zoom_factor > self.max_zoom) or \
+           (delta < 0 and current_transform.m11() * zoom_factor < self.min_zoom):
+            return
+            
+        self.scale(zoom_factor, zoom_factor)
+
+    def mousePressEvent(self, event):
+        if self.dragMode() == QGraphicsView.ScrollHandDrag:
+            self.viewport().setCursor(Qt.ClosedHandCursor)
+        super(CustomGraphicsView, self).mousePressEvent(event)
+
+    def mouseMoveEvent(self, event):
+        super(CustomGraphicsView, self).mouseMoveEvent(event)
 
     def mouseReleaseEvent(self, event):
-        if not self.editMode and event.button() == Qt.LeftButton and self.is_panning:
-            self.is_panning = False
-            self.setCursor(Qt.OpenHandCursor if self.dragMode() == QGraphicsView.ScrollHandDrag else Qt.ArrowCursor)
-            event.accept()
-        else:
-            super(CustomGraphicsView, self).mouseReleaseEvent(event)
-            
+        if self.dragMode() == QGraphicsView.ScrollHandDrag:
+            self.viewport().setCursor(Qt.OpenHandCursor)
+        super(CustomGraphicsView, self).mouseReleaseEvent(event)
+
     def event(self, event):
-        """Handle all events including touch gestures"""
-        if event.type() == QEvent.Gesture:
-            try:
-                return self.handleGestureEvent(event)
-            except Exception as e:
-                print(f"Gesture handling error: {e}")
-                
+        if self.isMacOS and event.type() == QEvent.Gesture:
+            return self.handleGestureEvent(event.gesture(Qt.PinchGesture))
         return super(CustomGraphicsView, self).event(event)
-    
-    def handleGestureEvent(self, event):
-        """Handle touch gesture events for zooming on macOS"""
-        try:
-            if hasattr(Qt, 'PinchGesture'):
-                pinch_gesture = event.gesture(Qt.PinchGesture)
-                if pinch_gesture:
-                    self.handlePinchGesture(pinch_gesture)
+
+    def handleGestureEvent(self, gesture):
+        if gesture:
+            if gesture.state() == Qt.GestureStarted:
+                self.gesture_started_transform = self.transform()
+            
+            if gesture.state() == Qt.GestureUpdated:
+                scale_factor = gesture.scaleFactor()
+                
+                # Prevent extreme zooming
+                if (self.gesture_started_transform.m11() * scale_factor > self.max_zoom) or \
+                   (self.gesture_started_transform.m11() * scale_factor < self.min_zoom):
                     return True
-        except Exception as e:
-            print(f"Error handling gesture: {e}")
+                
+                # Apply zoom
+                new_transform = QTransform(self.gesture_started_transform)
+                
+                # Get the center point of the gesture in scene coordinates
+                center_point = self.mapToScene(gesture.hotSpot().toPoint())
+                
+                new_transform.translate(center_point.x(), center_point.y())
+                new_transform.scale(scale_factor, scale_factor)
+                new_transform.translate(-center_point.x(), -center_point.y())
+                
+                self.setTransform(new_transform)
+
+            return True
         return False
-    
-    def handlePinchGesture(self, gesture):
-        """Handle pinch gesture for zooming"""
-        try:
-            if hasattr(gesture, 'state'):
-                if gesture.state() == Qt.GestureStarted:
-                    self.gestureInProgress = True
-                    self.lastGestureValue = gesture.scaleFactor()
-                elif gesture.state() == Qt.GestureUpdated and self.gestureInProgress:
-                    scale_change = gesture.scaleFactor() / self.lastGestureValue
-                    self.lastGestureValue = gesture.scaleFactor()
-                    
-                    current_scale = self.transform().m11()
-                    new_scale = current_scale * scale_change
-                    new_scale = max(self.min_zoom, min(self.max_zoom, new_scale))
-                    
-                    if current_scale != 0:
-                        scale_factor = new_scale / current_scale
-                        if abs(scale_factor - 1.0) > 1e-9:
-                            self.scale(scale_factor, scale_factor)
-                elif gesture.state() in [Qt.GestureFinished, Qt.GestureCanceled]:
-                    self.gestureInProgress = False
-        except Exception as e:
-            print(f"Error in pinch gesture: {e}")
-            self.gestureInProgress = False
 
     def updateContentFittingZoom(self, sceneRect):
         if sceneRect.isEmpty():
             return
-
-        padded_rect = sceneRect.adjusted(-30, -30, 30, 30)
         
-        view_width = self.viewport().width()
-        view_height = self.viewport().height()
+        viewRect = self.viewport().rect()
         
-        if view_width <= 0 or view_height <= 0:
+        if viewRect.width() <= 0 or viewRect.height() <= 0:
             return
             
-        scale_x = view_width / padded_rect.width()
-        scale_y = view_height / padded_rect.height()
+        x_ratio = viewRect.width() / sceneRect.width()
+        y_ratio = viewRect.height() / sceneRect.height()
         
-        fitting_scale = min(scale_x, scale_y)
-        
-        self.contentFittingZoom = max(0.01, fitting_scale * 0.9)
+        self.contentFittingZoom = min(x_ratio, y_ratio) * 0.95
 
 def create_sample_scene():
-    """Create a sample scene with editable items"""
     scene = CheckerboardGraphicsScene()
     
-    rect_item = QGraphicsRectItem(0, 0, 100, 60)
-    rect_item.setBrush(QBrush(QColor(255, 100, 100)))
+    rect_item = QGraphicsRectItem(0, 0, 100, 50)
+    rect_item.setBrush(QColor(255, 0, 0, 150))
+    rect_item.setPen(Qt.NoPen)
     rect_item.setPos(50, 50)
+    rect_item.setData(0, "rect1")
+    rect_item.setData(1, "Layer")
     scene.addItem(rect_item)
     
     ellipse_item = QGraphicsEllipseItem(0, 0, 80, 80)
-    ellipse_item.setBrush(QBrush(QColor(100, 255, 100)))
-    ellipse_item.setPos(200, 100)
+    ellipse_item.setBrush(QColor(0, 0, 255, 150))
+    ellipse_item.setPen(Qt.NoPen)
+    ellipse_item.setPos(200, 80)
+    ellipse_item.setData(0, "ellipse1")
+    ellipse_item.setData(1, "Layer")
     scene.addItem(ellipse_item)
     
-    scene.makeItemEditable(rect_item)
-    scene.makeItemEditable(ellipse_item)
+    scene.setEditMode(True)
     
     return scene
