+++ gui/exportoptions_window.py
@@ -8,6 +8,7 @@ class ExportOptionsDialog(QDialog):
         super().__init__(parent)
         self.choice = None
         self.config_manager = config_manager
+        self.parent = parent
 
         self.ui = Ui_ExportOptionsDialog()
         self.ui.setupUi(self)
@@ -26,13 +27,33 @@ class ExportOptionsDialog(QDialog):
         self.ui.tendiesButton.clicked.connect(lambda: self.set_choice("tendies"))
         self.ui.nuggetButton.clicked.connect(lambda: self.set_choice("nugget"))
 
-        # Dynamic icon based on parent's theme state (if available)
+        if parent and hasattr(parent, 'ca_files_in_tendies') and len(getattr(parent, 'ca_files_in_tendies', [])) > 1:
+            self.ui.copyButton.setText("Export as .ca files in folder")
+        
         is_dark = False
         if parent and hasattr(parent, 'isDarkMode'):
             is_dark = parent.isDarkMode
-        
-        nugget_icon_path = ":/icons/nugget-export-white.png" if is_dark else ":/icons/nugget-export.png"
-        self.ui.nuggetButton.setIcon(QIcon(nugget_icon_path))
+            
+            if is_dark:
+                self.ui.copyButton.setIcon(QIcon(":/icons/export-white.svg"))
+                self.ui.tendiesButton.setIcon(QIcon(":/icons/export-white.svg"))
+                nugget_icon_path = ":/icons/nugget-export-white.png" if is_dark else ":/icons/nugget-export.png"
+                self.ui.nuggetButton.setIcon(QIcon(nugget_icon_path))
+                
+                self.ui.titleLabel.setStyleSheet("font-size:18px; font-weight:bold; color: #FFFFFF;")
+                self.ui.copyButton.setStyleSheet("QToolButton { color: #FFFFFF; background-color: #3A3A3A; border: 1px solid #606060; border-radius: 8px; }")
+                self.ui.tendiesButton.setStyleSheet("QToolButton { color: #FFFFFF; background-color: #3A3A3A; border: 1px solid #606060; border-radius: 8px; }")
+                self.ui.nuggetButton.setStyleSheet("QToolButton { color: #FFFFFF; background-color: #3A3A3A; border: 1px solid #606060; border-radius: 8px; }")
+            else:
+                self.ui.copyButton.setIcon(QIcon(":/icons/export.svg"))
+                self.ui.tendiesButton.setIcon(QIcon(":/icons/export.svg"))
+                nugget_icon_path = ":/icons/nugget-export.png"
+                self.ui.nuggetButton.setIcon(QIcon(nugget_icon_path))
+
+                self.ui.titleLabel.setStyleSheet("font-size:18px; font-weight:bold; color: #303030;")
+                self.ui.copyButton.setStyleSheet("QToolButton { color: #303030; background-color: #F0F0F0; border: 1px solid #A0A0A0; border-radius: 8px; }")
+                self.ui.tendiesButton.setStyleSheet("QToolButton { color: #303030; background-color: #F0F0F0; border: 1px solid #A0A0A0; border-radius: 8px; }")
+                self.ui.nuggetButton.setStyleSheet("QToolButton { color: #303030; background-color: #F0F0F0; border: 1px solid #A0A0A0; border-radius: 8px; }")
 
     def set_choice(self, choice):
         self.choice = choice
