+++ themes/dark_style.qss
@@ -201,4 +201,88 @@ QTabBar::tab:selected {
 
 QTabBar::tab:hover:!selected {
     background-color: #4A4A4A;
+}
+
+/* Dialog styling */
+QDialog {
+    background-color: #303030;
+    color: #D0D0D0;
+}
+
+/* Input Dialog styling */
+QInputDialog {
+    background-color: #303030;
+    color: #D0D0D0;
+}
+
+QInputDialog QLabel {
+    color: #D0D0D0;
+}
+
+QInputDialog QLineEdit {
+    background-color: #3A3A3A;
+    color: #D0D0D0;
+    border: 1px solid #606060;
+    border-radius: 4px;
+    padding: 4px;
+}
+
+QInputDialog QPushButton {
+    border: 1px solid #606060;
+    border-radius: 6px;
+    padding: 5px 10px;
+    background-color: #3A3A3A;
+    color: #D0D0D0;
+}
+
+QInputDialog QPushButton:hover {
+    background-color: #4A4A4A;
+}
+
+QInputDialog QPushButton:pressed {
+    background-color: #5A5A5A;
+}
+
+QInputDialog QComboBox {
+    background-color: #3A3A3A;
+    color: #D0D0D0;
+    border: 1px solid #606060;
+    border-radius: 4px;
+    padding: 4px;
+}
+
+QInputDialog QListView {
+    background-color: #3A3A3A;
+    color: #D0D0D0;
+    border: 1px solid #606060;
+}
+
+QInputDialog QListView::item:selected {
+    background-color: #505050;
+    color: #FFFFFF;
+}
+
+/* Export Options Dialog styling */
+QDialog#ExportOptionsDialog {
+    background-color: #303030;
+    color: #D0D0D0;
+}
+
+QDialog#ExportOptionsDialog QLabel {
+    color: #D0D0D0;
+}
+
+QDialog#ExportOptionsDialog QToolButton {
+    background-color: #3A3A3A;
+    color: #D0D0D0;
+    border: 1px solid #606060;
+    border-radius: 8px;
+}
+
+QDialog#ExportOptionsDialog QToolButton:hover {
+    background-color: #4A4A4A;
+}
+
+QDialog#ExportOptionsDialog QToolButton:pressed {
+    background-color: #5A5A5A;
 }
