+++ themes/light_style.qss
@@ -202,4 +202,88 @@ QTabBar::tab:selected {
 
 QTabBar::tab:hover:!selected {
     background-color: #E0E0E0;
+}
+
+/* Dialog styling */
+QDialog {
+    background-color: #FFFFFF;
+    color: #303030;
+}
+
+/* Input Dialog styling */
+QInputDialog {
+    background-color: #FFFFFF;
+    color: #303030;
+}
+
+QInputDialog QLabel {
+    color: #303030;
+}
+
+QInputDialog QLineEdit {
+    background-color: #FFFFFF;
+    color: #303030;
+    border: 1px solid #A0A0A0;
+    border-radius: 4px;
+    padding: 4px;
+}
+
+QInputDialog QPushButton {
+    border: 1px solid #A0A0A0;
+    border-radius: 6px;
+    padding: 5px 10px;
+    background-color: #F0F0F0;
+    color: #303030;
+}
+
+QInputDialog QPushButton:hover {
+    background-color: #E0E0E0;
+}
+
+QInputDialog QPushButton:pressed {
+    background-color: #D0D0D0;
+}
+
+QInputDialog QComboBox {
+    background-color: #FFFFFF;
+    color: #303030;
+    border: 1px solid #A0A0A0;
+    border-radius: 4px;
+    padding: 4px;
+}
+
+QInputDialog QListView {
+    background-color: #FFFFFF;
+    color: #303030;
+    border: 1px solid #A0A0A0;
+}
+
+QInputDialog QListView::item:selected {
+    background-color: #E0E0E0;
+    color: #000000;
+}
+
+/* Export Options Dialog styling */
+QDialog#ExportOptionsDialog {
+    background-color: #FFFFFF;
+    color: #303030;
+}
+
+QDialog#ExportOptionsDialog QLabel {
+    color: #303030;
+}
+
+QDialog#ExportOptionsDialog QToolButton {
+    background-color: #F0F0F0;
+    color: #303030;
+    border: 1px solid #A0A0A0;
+    border-radius: 8px;
+}
+
+QDialog#ExportOptionsDialog QToolButton:hover {
+    background-color: #E0E0E0;
+}
+
+QDialog#ExportOptionsDialog QToolButton:pressed {
+    background-color: #D0D0D0;
 }
